chore: fix widget target
