+++ YRZYQuizWidget/Assets.xcassets/AccentColor.colorset/Contents.json
