+++ YRZYQuizWidget/Assets.xcassets/Contents.json
