+++ YRZYQuizWidget/Assets.xcassets/WidgetBackground.colorset/Contents.json
