+++ YRZYQuizWidget/YRZYQuizWidgetDebug.entitlements
