--- YRZYQuizWidgetExtension/YRZYQuizWidgetExtension.swift
@@ -1,69 +1,0 @@
-//
-//  YRZYQuizWidgetExtension.swift
-//  YRZYQuizWidgetExtension
-//
-//  Created by Shunya Yamada on 2021/05/16.
-//
-
-import WidgetKit
-import SwiftUI
-import Intents
-
-struct Provider: IntentTimelineProvider {
-    func placeholder(in context: Context) -> SimpleEntry {
-        SimpleEntry(date: Date(), configuration: ConfigurationIntent())
-    }
-
-    func getSnapshot(for configuration: ConfigurationIntent, in context: Context, completion: @escaping (SimpleEntry) -> ()) {
-        let entry = SimpleEntry(date: Date(), configuration: configuration)
-        completion(entry)
-    }
-
-    func getTimeline(for configuration: ConfigurationIntent, in context: Context, completion: @escaping (Timeline<Entry>) -> ()) {
-        var entries: [SimpleEntry] = []
-
-        // Generate a timeline consisting of five entries an hour apart, starting from the current date.
-        let currentDate = Date()
-        for hourOffset in 0 ..< 5 {
-            let entryDate = Calendar.current.date(byAdding: .hour, value: hourOffset, to: currentDate)!
-            let entry = SimpleEntry(date: entryDate, configuration: configuration)
-            entries.append(entry)
-        }
-
-        let timeline = Timeline(entries: entries, policy: .atEnd)
-        completion(timeline)
-    }
-}
-
-struct SimpleEntry: TimelineEntry {
-    let date: Date
-    let configuration: ConfigurationIntent
-}
-
-struct YRZYQuizWidgetExtensionEntryView : View {
-    var entry: Provider.Entry
-
-    var body: some View {
-        Text(entry.date, style: .time)
-    }
-}
-
-@main
-struct YRZYQuizWidgetExtension: Widget {
-    let kind: String = "YRZYQuizWidgetExtension"
-
-    var body: some WidgetConfiguration {
-        IntentConfiguration(kind: kind, intent: ConfigurationIntent.self, provider: Provider()) { entry in
-            YRZYQuizWidgetExtensionEntryView(entry: entry)
-        }
-        .configurationDisplayName("クイズの結果")
-        .description("今まで答えたクイズの結果を確認することができます。")
-    }
-}
-
-struct YRZYQuizWidgetExtension_Previews: PreviewProvider {
-    static var previews: some View {
-        YRZYQuizWidgetExtensionEntryView(entry: SimpleEntry(date: Date(), configuration: ConfigurationIntent()))
-            .previewContext(WidgetPreviewContext(family: .systemSmall))
-    }
-}
